Improve .env File Organization, Readability, and Documentation
This pull request aims to enhance the organization, readability, and understanding of the .env.template file for users when they modify the settings. The changes include organizing the file in a tree-like structure with appropriate comments, providing clear guidance for users about the purpose of each variable, their possible values, and default settings when applicable.

As a user with no prior knowledge of best practices of contributing to a project / .env.template file documentation, I took the liberty to make changes to the file based on what I would have liked to have seen when I first encountered it. My goal was to include every configurable option for ease of use and better understanding of how the code works.

The key improvements made in this pull request are:

1. Grouping related variables under appropriate headers for better organization and ease of navigation.
2. Adding informative comments for each variable to help users understand their purpose and possible values.
3. Including default values in the comments to inform users of the consequences of not providing a specific value for a variable, allowing them to make 
    informed decisions when configuring the application.
4. Formatting the file consistently for better readability.

These changes will enhance user experience by simplifying the configuration process and reducing potential confusion. Users can quickly and easily configure the application without having to search through the code to determine default values or understand the relationship between various settings. Additionally, well-organized code and documentation can lead to fewer issues and misunderstandings, saving time for both users and maintainers of the project.

Please review these changes and let me know if you have any questions or suggestions for further improvement so I can make any necessary adjustments.